requirements.txt
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==23.1.2
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
ipykernel==6.24.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==8.3.0
jupyter-core==5.3.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-magic==0.4.24
pytz==2022.1
pyzmq==25.1.0
regex==2023.8.8
requests==2.25.1
s3transfer==0.6.1
SecretStorage==3.3.1
selenium==4.12.0
service-identity==18.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.4
soupsieve==2.4.1
ssh-import-id==5.11
stack-data==0.6.2
systemd-python==234
tornado==6.3.2
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.4
Twisted==22.1.0
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.6
webdriver-manager==4.0.0
wsproto==1.2.0
xlrd==2.0.1
zipp==1.0.0
zope.interface==5.4.0
